adding new statistics

+++ desktop/Database.sql
@@ -71,7 +71,7 @@ CREATE TABLE IF NOT EXISTS Stock(
 	ID INT AUTO_INCREMENT not null PRIMARY KEY,
 	Name text,
     Amount int not null,
-    Price float not null,
+    Price double not null,
     Availability bool DEFAULT FALSE,
     DepartmentID INT NOT NULL,
